requirements.txt
astroid==2.3.3
certifi==2019.11.28
cycler==0.10.0
easydict==1.9
isort==4.3.21
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
matplotlib==2.2.4
mccabe==0.6.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.18.1
olefile==0.46
pandas==1.0.1
Pillow==7.0.0
pylint==2.4.4
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
scipy==1.4.1
six==1.14.0
torch==1.4.0
torchvision==0.5.0
tornado==6.0.3
typed-ast==1.4.1
wrapt==1.11.2
